Move card detection into seperate file. Rename detect_card.cpp and .h
