+++ include/DetectCard.h
