--- src/detect_card.cpp
@@ -1,9 +1,0 @@
-#include "detect_card.h"
-
-DetectCard::DetectCard(/* args */)
-{
-}
-
-DetectCard::~DetectCard()
-{
-}
